Try to bump versions incrementally
