+++ build-logic/gradle/wrapper/gradle-wrapper.properties
@@ -15,7 +15,7 @@
 #
 #Sun Dec 05 12:14:24 EST 2021
 distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.9-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.11.1-bin.zip
 distributionPath=wrapper/dists
 zipStorePath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
