+++ common/ui/archive-detail/src/commonMain/kotlin/com/tunjid/me/archivedetail/ArchiveDetailRoute.kt
@@ -44,7 +44,6 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.tunjid.me.archivedetail.BlogMarkdownScope.Companion.BlogMarkdownScope
 import com.tunjid.me.archivedetail.di.kind
 import com.tunjid.me.core.model.Descriptor
 import com.tunjid.me.core.ui.AsyncRasterImage
@@ -78,101 +77,101 @@ internal fun ArchiveDetailScreen(
     state: State,
     actions: (Action) -> Unit,
     modifier: Modifier = Modifier,
-) = BlogMarkdownScope {
-    val scrollState = rememberLazyListState()
-    val navBarSizeDp = with(LocalDensity.current) { state.navBarSize.toDp() }
+) = BlogMarkdown(
+    modifier = modifier,
+    markdown = state.archive?.body ?: "",
+    content = { markdownState, components, innerModifier ->
+        val scrollState = rememberLazyListState()
+        val navBarSizeDp = with(LocalDensity.current) { state.navBarSize.toDp() }
 
-    val archive = state.archive
-
-    // Close the secondary pane when invoking back since it contains the list view
-    SecondaryPaneCloseBackHandler(
-        enabled = state.isInPrimaryNav && state.hasSecondaryPanel
-    )
-
-    val markdownState = innerMarkdownState(archive?.body ?: "")
-
-    LazyColumn(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally,
-        state = scrollState
-    ) {
-        stickyHeader {
-            TopAppBar(
-                state = state,
-                actions = actions,
-            )
-        }
-        item {
-            movableSharedElementScope.updatedMovableSharedElementOf(
-                key = state.sharedElementKey,
-                MediaArgs(
-                    url = state.headerThumbnail,
-                    contentScale = ContentScale.Crop,
-                ),
-                modifier = Modifier
-                    .heightIn(max = 300.dp)
-                    .aspectRatio(ratio = 16f / 9f)
-                    .padding(horizontal = 16.dp)
-                    .clip(MaterialTheme.shapes.medium),
-                sharedElement = { state, innerModifier ->
-                    AsyncRasterImage(
-                        args = state,
-                        modifier = innerModifier
-                    )
-                }
-            )
-        }
-        item {
-            Chips(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-                name = "Categories:",
-                chipInfoList = state.descriptorChips<Descriptor.Category>(),
-            )
-        }
-
-        item {
-            Spacer(modifier = Modifier.padding(16.dp))
-        }
-
-        blogItems(
-            lazyListScope = this,
-            innerMarkdownState = markdownState,
+        // Close the secondary pane when invoking back since it contains the list view
+        SecondaryPaneCloseBackHandler(
+            enabled = state.isInPrimaryNav && state.hasSecondaryPanel
         )
 
-        item {
-            Spacer(modifier = Modifier.padding(16.dp))
-        }
+        LazyColumn(
+            modifier = innerModifier,
+            horizontalAlignment = Alignment.CenterHorizontally,
+            state = scrollState
+        ) {
+            stickyHeader {
+                TopAppBar(
+                    state = state,
+                    actions = actions,
+                )
+            }
+            item {
+                movableSharedElementScope.updatedMovableSharedElementOf(
+                    key = state.sharedElementKey,
+                    MediaArgs(
+                        url = state.headerThumbnail,
+                        contentScale = ContentScale.Crop,
+                    ),
+                    modifier = Modifier
+                        .heightIn(max = 300.dp)
+                        .aspectRatio(ratio = 16f / 9f)
+                        .padding(horizontal = 16.dp)
+                        .clip(MaterialTheme.shapes.medium),
+                    sharedElement = { state, innerModifier ->
+                        AsyncRasterImage(
+                            args = state,
+                            modifier = innerModifier
+                        )
+                    }
+                )
+            }
+            item {
+                Chips(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = 16.dp),
+                    name = "Categories:",
+                    chipInfoList = state.descriptorChips<Descriptor.Category>(),
+                )
+            }
 
-        item {
-            Chips(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-                name = "Tags:",
-                chipInfoList = state.descriptorChips<Descriptor.Tag>(),
+            item {
+                Spacer(modifier = Modifier.padding(16.dp))
+            }
+
+            blogItems(
+                innerMarkdownState = markdownState,
+                components = components,
             )
+
+            item {
+                Spacer(modifier = Modifier.padding(16.dp))
+            }
+
+            item {
+                Chips(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = 16.dp),
+                    name = "Tags:",
+                    chipInfoList = state.descriptorChips<Descriptor.Tag>(),
+                )
+            }
+
+            item {
+                Spacer(modifier = Modifier.padding(64.dp + navBarSizeDp))
+            }
         }
 
-        item {
-            Spacer(modifier = Modifier.padding(64.dp + navBarSizeDp))
+        // Pop nav if this archive does not exist anymore
+        val wasDeleted = state.wasDeleted
+        LaunchedEffect(wasDeleted) {
+            if (wasDeleted) actions(Action.Navigate.Pop)
+        }
+
+        // If the user fully expands the secondary pane, pop this destination back to the list
+        LaunchedEffect(state.hasSecondaryPanel, state.paneAnchor) {
+            if (state.hasSecondaryPanel && state.paneAnchor == PaneAnchor.Full) {
+                actions(Action.Navigate.Pop)
+            }
         }
     }
-
-    // Pop nav if this archive does not exist anymore
-    val wasDeleted = state.wasDeleted
-    LaunchedEffect(wasDeleted) {
-        if (wasDeleted) actions(Action.Navigate.Pop)
-    }
-
-    // If the user fully expands the secondary pane, pop this destination back to the list
-    LaunchedEffect(state.hasSecondaryPanel, state.paneAnchor) {
-        if (state.hasSecondaryPanel && state.paneAnchor == PaneAnchor.Full) {
-            actions(Action.Navigate.Pop)
-        }
-    }
-}
+)
 
 @Composable
 private fun TopAppBar(
