+++ gradle/libs.versions.toml
@@ -2,7 +2,7 @@
 accompanist = "0.28.0"
 apacheCommonsImaging = "1.0-alpha3"
 androidDesugarJdkLibs = "2.1.5"
-androidGradlePlugin = "8.5.2"
+androidGradlePlugin = "8.9.1"
 androidxActivity = "1.10.1"
 androidxAppCompat = "1.7.0"
 androidxCore = "1.16.0"
@@ -16,6 +16,7 @@ androidxEspresso = "3.6.1"
 androidxTest = "1.6.1"
 androidxTestExt = "1.2.1"
 androidxWork = "2.10.0"
+compose-material-icons = "1.7.3"
 googleMaterial = "1.12.0"
 gradleKotlinDsl = "5.2.0"
 jakeWhartonRetrofit = "1.0.0"
@@ -30,11 +31,11 @@ imageIO = "3.9.4"
 ksp = "2.1.20-1.0.32"
 ktor = "3.0.3"
 mockk = "1.12.3"
-multiplatformMarkdownRenderer = "0.33.0-rc01"
+multiplatformMarkdownRenderer = "0.33.0"
 okhttp = "4.12.0"
 okio = "3.10.2"
 retrofit = "2.11.0"
-richtext = "0.16.0"
+richtext = "1.0.0-alpha02"
 richeditor = "1.0.0-rc12"
 socketio = "2.0.0"
 sqldelight = "2.0.0"
@@ -97,6 +98,7 @@ jetbrains-compose-animation = { group = "org.jetbrains.compose.animation", name 
 jetbrains-compose-foundation-foundation = { group = "org.jetbrains.compose.foundation", name = "foundation", version.ref = "jetbrainsCompose" }
 jetbrains-compose-foundation-layout = { group = "org.jetbrains.compose.foundation", name = "foundation-layout", version.ref = "jetbrainsCompose" }
 jetbrains-compose-material = { group = "org.jetbrains.compose.material", name = "material", version.ref = "jetbrainsCompose" }
+jetbrains-compose-material-icons-extended = { group = "org.jetbrains.compose.material", name = "material-icons-extended", version.ref = "compose-material-icons" }
 jetbrains-compose-material3 = { group = "org.jetbrains.compose.material3", name = "material3", version.ref = "jetbrainsCompose" }
 jetbrains-compose-gradlePlugin = { group = "org.jetbrains.compose", name = "compose-gradle-plugin", version.ref = "jetbrainsCompose" }
 jetbrains-compose-runtime = { group = "org.jetbrains.compose.runtime", name = "runtime", version.ref = "jetbrainsCompose" }
